requirements.txt
faker
datetime
random
confluent_kafka
python-dotenv
